AI/microservices/receipt-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
requests==2.32.3
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.0.0
# Default CPU runtime. Override with a matching paddlepaddle-gpu package when GPU support is available.
paddlepaddle==3.3.1
paddleocr==3.4.0
